requirements.txt
Django>=4.1.6, <4.1.7
djangorestframework>=3.14.0, <3.14.1
psycopg2>=2.9.5, <3.0
Pillow>=9.4.0, <9.4.1
flake8>=5.0.4, <5.0.5
django-debug-toolbar>=3.8.1, <3.8.2